fix collision

+++ src/draw/draw_floor_and_cealing.c
@@ -20,7 +20,7 @@ uint32_t blend_color(uint32_t originalColor, float blendFactor) {
     r *= (1 - blendFactor);
     g *= (1 - blendFactor);
     b *= (1 - blendFactor);
-    return (r << 24) | (g << 16) | (b << 8) | 0xFF; // Assuming alpha is always 255
+    return ((uint32_t)r << 24) | ((uint32_t)g << 16) | ((uint32_t)b << 8) | 0xFF; // Assuming alpha is always 255
 }
 
 uint32_t blend_two_colors(uint32_t color1, uint32_t color2, float blendFactor) {
@@ -67,10 +67,6 @@ inline void draw_floor_and_ceiling(t_data *data, int y) {
         float blendFactor = fminf(1.0, fmaxf(0.0, 1.0 - position / half_height));
 
         uint32_t floorColor = blend_color(get_pixel(data->texture[FLOOR], texture.x, texture.y), blendFactor);
-
-        // if (is_in_circle((t_vector_2d){20,12}, 0.5, (t_vector_2d){floor.x, floor.y})) {
-        //     floorColor = blend_two_colors(floorColor, ((uint32_t)255 << 24) | ((uint32_t)255 << 16) | ((uint32_t)0 << 8) | 0xFF, 0.1);
-        // }
         uint32_t ceilingColor = blend_color(get_pixel(data->texture[CEILING], texture.x, texture.y), blendFactor);
 
         mlx_put_pixel(data->window->image, x, y, floorColor);
